frontend: move repo functions to a separate file
To avoid circular dependencies when we introduce some database queries
to them.
